Rename TS EnumXXX examples to VariantXXX.
