+++ README.md
@@ -338,7 +338,7 @@ Boolean/integer/float constants can be expressed as ``` | [@genType.as true] Tru
 The `@genType.as` annotation can also be used on variants with payloads to determine what appears in `{ tag: ... }`.
 
 
-For more examples, see [Enums.re](examples/typescript-react-example/src/Enums.re) and [EnumsWithPayload.re](examples/typescript-react-example/src/EnumsWithPayload.re).
+For more examples, see [Variants.re](examples/typescript-react-example/src/Variants.re) and [VariantsWithPayload.re](examples/typescript-react-example/src/VariantsWithPayload.re).
 
 **NOTE** When exporting/importing values that have polymorphic variant type, you have to use type annotations, and cannot rely on type inference. So instead of ```let monday = `Monday```, use ```let monday : days = `Monday```. The former does not work, as the type checker infers a tyoe without annotations.
