--- examples/typescript-react-example/src/Enums.bs.js
@@ -1,75 +1,0 @@
-// Generated by BUCKLESCRIPT VERSION 4.0.14, PLEASE EDIT WITH CARE
-
-
-function isWeekend(x) {
-  if (x !== -29784519) {
-    return x === 569248848;
-  } else {
-    return true;
-  }
-}
-
-function onlySunday(param) {
-  return /* () */0;
-}
-
-function swap(x) {
-  if (x >= 569248848) {
-    return /* saturday */-29784519;
-  } else {
-    return /* sunday */569248848;
-  }
-}
-
-function testConvert(x) {
-  return x;
-}
-
-function testConvert2(x) {
-  return x;
-}
-
-function testConvert3(x) {
-  return x;
-}
-
-function testConvert2to3(x) {
-  return x;
-}
-
-function id1(x) {
-  return x;
-}
-
-function id2(x) {
-  return x;
-}
-
-var monday = /* monday */-949852400;
-
-var saturday = /* saturday */-29784519;
-
-var sunday = /* sunday */569248848;
-
-var fortytwoOK = /* fortytwo */23437694;
-
-var fortytwoBAD = /* fortytwo */23437694;
-
-export {
-  isWeekend ,
-  monday ,
-  saturday ,
-  sunday ,
-  onlySunday ,
-  swap ,
-  testConvert ,
-  fortytwoOK ,
-  fortytwoBAD ,
-  testConvert2 ,
-  testConvert3 ,
-  testConvert2to3 ,
-  id1 ,
-  id2 ,
-  
-}
-/* No side effect */
